System test cases for lobster-report qualification
- System test cases to qualify lobster-report tool
- Fixed imports for python files in lobster-report
- Added more dummy requirements for testing purpose
