+++ tests_system/lobster_report/data/code_linear.lobster
@@ -1,0 +1,102 @@
+{
+  "data": [
+    {
+      "tag": "python simple_book_system.SimpleBookSystem.search_books",
+      "location": {
+        "kind": "file",
+        "file": "simple_book_system.py",
+        "line": 24,
+        "column": null
+      },
+      "name": "simple_book_system.SimpleBookSystem.search_books",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req software_example.search_books"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python simple_book_system.SimpleBookSystem.track_progress",
+      "location": {
+        "kind": "file",
+        "file": "simple_book_system.py",
+        "line": 41,
+        "column": null
+      },
+      "name": "simple_book_system.SimpleBookSystem.track_progress",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req software_example.reading_tracker"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python simple_book_system.SimpleBookSystem.organize_library",
+      "location": {
+        "kind": "file",
+        "file": "simple_book_system.py",
+        "line": 59,
+        "column": null
+      },
+      "name": "simple_book_system.SimpleBookSystem.organize_library",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req software_example.library_manager"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python simple_book_system.SimpleBookSystem.add_review",
+      "location": {
+        "kind": "file",
+        "file": "simple_book_system.py",
+        "line": 74,
+        "column": null
+      },
+      "name": "simple_book_system.SimpleBookSystem.add_review",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req software_example.book_reviews"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python simple_book_system.SimpleBookSystem.show_reviews",
+      "location": {
+        "kind": "file",
+        "file": "simple_book_system.py",
+        "line": 90,
+        "column": null
+      },
+      "name": "simple_book_system.SimpleBookSystem.show_reviews",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req software_example.book_reviews"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    }
+  ],
+  "generator": "lobster_python",
+  "schema": "lobster-imp-trace",
+  "version": 3
+}
