+++ tests_system/lobster_report/data/five_level_policy.conf
@@ -1,0 +1,22 @@
+requirements "System Requirements" {
+  source: "system_requirements.lobster";
+}
+
+requirements "Software Requirements" {
+  source: "software_requirements.lobster";
+}
+
+implementation "Code" {
+  source: "code.lobster";
+  trace to: "Software Requirements";
+}
+
+activity "System Test" {
+  source: "system-tests.lobster";
+  trace to: "System Requirements";
+}
+
+activity "Software Test" {
+  source: "unit-tests.lobster";
+  trace to: "Software Requirements";
+}
