+++ tests_system/lobster_report/data/linear_policy.conf
@@ -1,0 +1,13 @@
+requirements "System Requirements" {
+    source: "linear_system_requirements.lobster";
+}
+
+requirements "Software Requirements" {
+    source: "linear_software_requirements.lobster";
+    trace to: "System Requirements";
+}
+
+implementation "Code" {
+    source: "code_linear.lobster";
+    trace to: "Software Requirements";
+}
