+++ tests_system/lobster_report/data/linear_software_requirements.lobster
@@ -1,0 +1,91 @@
+{
+  "data": [
+    {
+      "tag": "req software_example.search_books",
+      "location": {
+        "kind": "file",
+        "file": "linear_requirements.trlc",
+        "line": 20,
+        "column": 30
+      },
+      "name": "software_example.search_books",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req software_example.finder"
+      ],
+      "framework": "TRLC",
+      "kind": "software_requirement",
+      "text": "The software shall implement book search functionality",
+      "status": null
+    },
+    {
+      "tag": "req software_example.reading_tracker",
+      "location": {
+        "kind": "file",
+        "file": "linear_requirements.trlc",
+        "line": 25,
+        "column": 30
+      },
+      "name": "software_example.reading_tracker",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req software_example.tracker"
+      ],
+      "framework": "TRLC",
+      "kind": "software_requirement",
+      "text": "The software shall track reading progress and statistics",
+      "status": null
+    },
+    {
+      "tag": "req software_example.library_manager",
+      "location": {
+        "kind": "file",
+        "file": "linear_requirements.trlc",
+        "line": 30,
+        "column": 30
+      },
+      "name": "software_example.library_manager",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req software_example.organizer"
+      ],
+      "framework": "TRLC",
+      "kind": "software_requirement",
+      "text": "The software shall organize and categorize books",
+      "status": null
+    },
+    {
+      "tag": "req software_example.book_reviews",
+      "location": {
+        "kind": "file",
+        "file": "linear_requirements.trlc",
+        "line": 35,
+        "column": 30
+      },
+      "name": "software_example.book_reviews",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req software_example.reviews"
+      ],
+      "framework": "TRLC",
+      "kind": "software_requirement",
+      "text": "The software shall collect and display book reviews",
+      "status": null
+    }
+  ],
+  "generator": "lobster-trlc",
+  "schema": "lobster-req-trace",
+  "version": 4
+}
