+++ tests_system/lobster_report/data/linear_system_requirements.lobster
@@ -1,0 +1,79 @@
+{
+  "data": [
+    {
+      "tag": "req software_example.finder",
+      "location": {
+        "kind": "file",
+        "file": "linear_requirements.trlc",
+        "line": 4,
+        "column": 28
+      },
+      "name": "software_example.finder",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "system_requirement",
+      "text": "The system shall find books that people actually want to read",
+      "status": null
+    },
+    {
+      "tag": "req software_example.tracker",
+      "location": {
+        "kind": "file",
+        "file": "linear_requirements.trlc",
+        "line": 8,
+        "column": 28
+      },
+      "name": "software_example.tracker",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "system_requirement",
+      "text": "The system shall count pages read without making people feel guilty",
+      "status": null
+    },
+    {
+      "tag": "req software_example.organizer",
+      "location": {
+        "kind": "file",
+        "file": "linear_requirements.trlc",
+        "line": 12,
+        "column": 28
+      },
+      "name": "software_example.organizer",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "system_requirement",
+      "text": "The system shall arrange books so they can be found again",
+      "status": null
+    },
+    {
+      "tag": "req software_example.reviews",
+      "location": {
+        "kind": "file",
+        "file": "linear_requirements.trlc",
+        "line": 16,
+        "column": 28
+      },
+      "name": "software_example.reviews",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "system_requirement",
+      "text": "The system shall gather opinions about books from real humans",
+      "status": null
+    }
+  ],
+  "generator": "lobster-trlc",
+  "schema": "lobster-req-trace",
+  "version": 4
+}
