+++ tests_system/lobster_report/data/multiple_traces_code.lobster
@@ -4,7 +4,7 @@
       "tag": "python vehicle_implementation.StupidVehicle.vehicle_exists",
       "location": {
         "kind": "file",
-        "file": ".\\vehicle_implementation.py",
+        "file": "vehicle_implementation.py",
         "line": 24,
         "column": null
       },
@@ -23,7 +23,7 @@
       "tag": "python vehicle_implementation.StupidVehicle.run_engine",
       "location": {
         "kind": "file",
-        "file": ".\\vehicle_implementation.py",
+        "file": "vehicle_implementation.py",
         "line": 30,
         "column": null
       },
@@ -42,7 +42,7 @@
       "tag": "python vehicle_implementation.StupidVehicle.activate_safety_system",
       "location": {
         "kind": "file",
-        "file": ".\\vehicle_implementation.py",
+        "file": "vehicle_implementation.py",
         "line": 43,
         "column": null
       },
@@ -61,7 +61,7 @@
       "tag": "python vehicle_implementation.StupidVehicle.apply_brakes",
       "location": {
         "kind": "file",
-        "file": ".\\vehicle_implementation.py",
+        "file": "vehicle_implementation.py",
         "line": 63,
         "column": null
       },
@@ -80,7 +80,7 @@
       "tag": "python vehicle_implementation.StupidVehicle.check_electrical_system",
       "location": {
         "kind": "file",
-        "file": ".\\vehicle_implementation.py",
+        "file": "vehicle_implementation.py",
         "line": 84,
         "column": null
       },
@@ -99,7 +99,7 @@
       "tag": "python vehicle_implementation.StupidVehicle.operate_fuel_system",
       "location": {
         "kind": "file",
-        "file": ".\\vehicle_implementation.py",
+        "file": "vehicle_implementation.py",
         "line": 101,
         "column": null
       },
@@ -118,7 +118,7 @@
       "tag": "python vehicle_implementation.StupidVehicle.operate_transmission_system",
       "location": {
         "kind": "file",
-        "file": ".\\vehicle_implementation.py",
+        "file": "vehicle_implementation.py",
         "line": 115,
         "column": null
       },
@@ -137,7 +137,7 @@
       "tag": "python vehicle_implementation.StupidVehicle.control_steering_system",
       "location": {
         "kind": "file",
-        "file": ".\\vehicle_implementation.py",
+        "file": "vehicle_implementation.py",
         "line": 131,
         "column": null
       },
@@ -156,7 +156,7 @@
       "tag": "python vehicle_implementation.StupidVehicle.operate_lighting_system",
       "location": {
         "kind": "file",
-        "file": ".\\vehicle_implementation.py",
+        "file": "vehicle_implementation.py",
         "line": 150,
         "column": null
       },
@@ -175,7 +175,7 @@
       "tag": "python vehicle_implementation.StupidVehicle.control_climate_system",
       "location": {
         "kind": "file",
-        "file": ".\\vehicle_implementation.py",
+        "file": "vehicle_implementation.py",
         "line": 176,
         "column": null
       },
@@ -186,6 +186,49 @@
       "just_global": [],
       "refs": [
         "req example.climate_control"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python vehicle_implementation.StupidVehicle.comprehensive_system_check",
+      "location": {
+        "kind": "file",
+        "file": "vehicle_implementation.py",
+        "line": 203,
+        "column": null
+      },
+      "name": "vehicle_implementation.StupidVehicle.comprehensive_system_check",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req example.engine",
+        "req example.electrical_system",
+        "req example.fuel_system",
+        "req example.safety_system"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python vehicle_implementation.StupidVehicle.run_vehicle_diagnostics",
+      "location": {
+        "kind": "file",
+        "file": "vehicle_implementation.py",
+        "line": 225,
+        "column": null
+      },
+      "name": "vehicle_implementation.StupidVehicle.run_vehicle_diagnostics",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req example.vehicle",
+        "req example.engine",
+        "req example.safety_system"
       ],
       "language": "Python",
       "kind": "Method"
