+++ tests_system/lobster_report/data/multiple_traces_just.conf
@@ -1,0 +1,13 @@
+requirements "Requirements" {
+    source: "just_requirements.lobster";
+}
+
+implementation "Code" {
+    source: "multiple_traces_code.lobster";
+    trace to: "Requirements";
+}
+
+activity "Test" {
+    source: "multiple_traces_test.lobster";
+    trace to: "Requirements";
+}
