+++ tests_system/lobster_report/data/pizza_code.lobster
@@ -1,0 +1,267 @@
+{
+  "data": [
+    {
+      "tag": "python pizza_implementation.OvenController.start_cooking",
+      "location": {
+        "kind": "file",
+        "file": "pizza_implementation.py",
+        "line": 37,
+        "column": null
+      },
+      "name": "pizza_implementation.OvenController.start_cooking",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.oven_controller",
+        "req pizza_example.oven"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python pizza_implementation.OvenController.check_cooking_status",
+      "location": {
+        "kind": "file",
+        "file": "pizza_implementation.py",
+        "line": 45,
+        "column": null
+      },
+      "name": "pizza_implementation.OvenController.check_cooking_status",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.oven_controller",
+        "req pizza_example.oven"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python pizza_implementation.RouteOptimizer.calculate_route",
+      "location": {
+        "kind": "file",
+        "file": "pizza_implementation.py",
+        "line": 66,
+        "column": null
+      },
+      "name": "pizza_implementation.RouteOptimizer.calculate_route",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.route_planner",
+        "req pizza_example.delivery"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python pizza_implementation.OrderProcessor.create_customer",
+      "location": {
+        "kind": "file",
+        "file": "pizza_implementation.py",
+        "line": 95,
+        "column": null
+      },
+      "name": "pizza_implementation.OrderProcessor.create_customer",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.order_processor",
+        "req pizza_example.ordering"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python pizza_implementation.OrderProcessor.place_order",
+      "location": {
+        "kind": "file",
+        "file": "pizza_implementation.py",
+        "line": 105,
+        "column": null
+      },
+      "name": "pizza_implementation.OrderProcessor.place_order",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.order_processor",
+        "req pizza_example.ordering"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python pizza_implementation.PaymentGateway.process_payment",
+      "location": {
+        "kind": "file",
+        "file": "pizza_implementation.py",
+        "line": 122,
+        "column": null
+      },
+      "name": "pizza_implementation.PaymentGateway.process_payment",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.payment_gateway",
+        "req pizza_example.payment"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python pizza_implementation.DeliveryTracker.start_delivery",
+      "location": {
+        "kind": "file",
+        "file": "pizza_implementation.py",
+        "line": 150,
+        "column": null
+      },
+      "name": "pizza_implementation.DeliveryTracker.start_delivery",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.gps_tracker",
+        "req pizza_example.tracking"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python pizza_implementation.InventoryManager.use_ingredients",
+      "location": {
+        "kind": "file",
+        "file": "pizza_implementation.py",
+        "line": 177,
+        "column": null
+      },
+      "name": "pizza_implementation.InventoryManager.use_ingredients",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.stock_manager",
+        "req pizza_example.inventory"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python pizza_implementation.InventoryManager.check_for_alerts",
+      "location": {
+        "kind": "file",
+        "file": "pizza_implementation.py",
+        "line": 195,
+        "column": null
+      },
+      "name": "pizza_implementation.InventoryManager.check_for_alerts",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.stock_manager",
+        "req pizza_example.inventory"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python pizza_implementation.NotificationSystem.send_sms",
+      "location": {
+        "kind": "file",
+        "file": "pizza_implementation.py",
+        "line": 213,
+        "column": null
+      },
+      "name": "pizza_implementation.NotificationSystem.send_sms",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.notification_system",
+        "req pizza_example.communication"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python pizza_implementation.NotificationSystem.notify_order_status",
+      "location": {
+        "kind": "file",
+        "file": "pizza_implementation.py",
+        "line": 226,
+        "column": null
+      },
+      "name": "pizza_implementation.NotificationSystem.notify_order_status",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.notification_system",
+        "req pizza_example.communication"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python pizza_implementation.RecipeDatabase.get_recipe",
+      "location": {
+        "kind": "file",
+        "file": "pizza_implementation.py",
+        "line": 258,
+        "column": null
+      },
+      "name": "pizza_implementation.RecipeDatabase.get_recipe",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.recipe_database",
+        "req pizza_example.quality"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python pizza_implementation.RecipeDatabase.validate_quality",
+      "location": {
+        "kind": "file",
+        "file": "pizza_implementation.py",
+        "line": 264,
+        "column": null
+      },
+      "name": "pizza_implementation.RecipeDatabase.validate_quality",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.recipe_database",
+        "req pizza_example.quality"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    }
+  ],
+  "generator": "lobster_python",
+  "schema": "lobster-imp-trace",
+  "version": 3
+}
