+++ tests_system/lobster_report/data/pizza_policy.conf
@@ -1,0 +1,26 @@
+requirements "System Requirements" {
+   source: "pizza_system_requirements.lobster";
+   requires: "Component Tests";
+}
+
+requirements "Software Requirements" {
+   source: "pizza_software_requirements.lobster";
+   requires: "Component Tests" or "Unit Tests";
+}
+
+implementation "Code" {
+   source: "pizza_code.lobster";
+   trace to: "System Requirements";
+   trace to: "Software Requirements";
+}
+
+activity "Component Tests" {
+   source: "pizza_component_tests.lobster";
+   trace to: "System Requirements";
+   trace to: "Software Requirements";
+}
+
+activity "Unit Tests" {
+   source: "pizza_unit_tests.lobster";
+   trace to: "Software Requirements";
+}
