+++ tests_system/lobster_report/data/pizza_software_requirements.lobster
@@ -1,0 +1,151 @@
+{
+  "data": [
+    {
+      "tag": "req pizza_example.oven_controller",
+      "location": {
+        "kind": "file",
+        "file": "pizza_software_requirements.trlc",
+        "line": 4,
+        "column": 30
+      },
+      "name": "pizza_example.oven_controller",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "software_requirement",
+      "text": "The software shall control oven temperature and cooking timer",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.route_planner",
+      "location": {
+        "kind": "file",
+        "file": "pizza_software_requirements.trlc",
+        "line": 8,
+        "column": 30
+      },
+      "name": "pizza_example.route_planner",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "software_requirement",
+      "text": "The software shall calculate fastest delivery routes",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.order_processor",
+      "location": {
+        "kind": "file",
+        "file": "pizza_software_requirements.trlc",
+        "line": 12,
+        "column": 30
+      },
+      "name": "pizza_example.order_processor",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "software_requirement",
+      "text": "The software shall manage pizza orders and customer details",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.payment_gateway",
+      "location": {
+        "kind": "file",
+        "file": "pizza_software_requirements.trlc",
+        "line": 16,
+        "column": 30
+      },
+      "name": "pizza_example.payment_gateway",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "software_requirement",
+      "text": "The software shall handle credit card transactions securely",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.gps_tracker",
+      "location": {
+        "kind": "file",
+        "file": "pizza_software_requirements.trlc",
+        "line": 20,
+        "column": 30
+      },
+      "name": "pizza_example.gps_tracker",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "software_requirement",
+      "text": "The software shall provide real-time delivery location updates",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.stock_manager",
+      "location": {
+        "kind": "file",
+        "file": "pizza_software_requirements.trlc",
+        "line": 24,
+        "column": 30
+      },
+      "name": "pizza_example.stock_manager",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "software_requirement",
+      "text": "The software shall monitor ingredient levels and send alerts",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.notification_system",
+      "location": {
+        "kind": "file",
+        "file": "pizza_software_requirements.trlc",
+        "line": 28,
+        "column": 30
+      },
+      "name": "pizza_example.notification_system",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "software_requirement",
+      "text": "The software shall send SMS and email updates to customers",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.recipe_database",
+      "location": {
+        "kind": "file",
+        "file": "pizza_software_requirements.trlc",
+        "line": 32,
+        "column": 30
+      },
+      "name": "pizza_example.recipe_database",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "software_requirement",
+      "text": "The software shall store pizza recipes and quality standards",
+      "status": null
+    }
+  ],
+  "generator": "lobster-trlc",
+  "schema": "lobster-req-trace",
+  "version": 4
+}
