+++ tests_system/lobster_report/data/pizza_system_requirements.lobster
@@ -1,0 +1,151 @@
+{
+  "data": [
+    {
+      "tag": "req pizza_example.oven",
+      "location": {
+        "kind": "file",
+        "file": "pizza_system_requirements.trlc",
+        "line": 4,
+        "column": 28
+      },
+      "name": "pizza_example.oven",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "system_requirement",
+      "text": "The system shall heat pizzas to perfect crispy goodness",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.delivery",
+      "location": {
+        "kind": "file",
+        "file": "pizza_system_requirements.trlc",
+        "line": 8,
+        "column": 28
+      },
+      "name": "pizza_example.delivery",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "system_requirement",
+      "text": "The system shall transport pizzas while they are still hot",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.ordering",
+      "location": {
+        "kind": "file",
+        "file": "pizza_system_requirements.trlc",
+        "line": 12,
+        "column": 28
+      },
+      "name": "pizza_example.ordering",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "system_requirement",
+      "text": "The system shall accept pizza orders from hungry customers",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.payment",
+      "location": {
+        "kind": "file",
+        "file": "pizza_system_requirements.trlc",
+        "line": 16,
+        "column": 28
+      },
+      "name": "pizza_example.payment",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "system_requirement",
+      "text": "The system shall process money without stealing tips",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.tracking",
+      "location": {
+        "kind": "file",
+        "file": "pizza_system_requirements.trlc",
+        "line": 20,
+        "column": 28
+      },
+      "name": "pizza_example.tracking",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "system_requirement",
+      "text": "The system shall show where the pizza is on its journey",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.inventory",
+      "location": {
+        "kind": "file",
+        "file": "pizza_system_requirements.trlc",
+        "line": 24,
+        "column": 28
+      },
+      "name": "pizza_example.inventory",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "system_requirement",
+      "text": "The system shall keep track of cheese and pepperoni supplies",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.communication",
+      "location": {
+        "kind": "file",
+        "file": "pizza_system_requirements.trlc",
+        "line": 28,
+        "column": 28
+      },
+      "name": "pizza_example.communication",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "system_requirement",
+      "text": "The system shall notify customers when pizza arrives",
+      "status": null
+    },
+    {
+      "tag": "req pizza_example.quality",
+      "location": {
+        "kind": "file",
+        "file": "pizza_system_requirements.trlc",
+        "line": 32,
+        "column": 28
+      },
+      "name": "pizza_example.quality",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "framework": "TRLC",
+      "kind": "system_requirement",
+      "text": "The system shall ensure pizzas taste better than cardboard",
+      "status": null
+    }
+  ],
+  "generator": "lobster-trlc",
+  "schema": "lobster-req-trace",
+  "version": 4
+}
