+++ tests_system/lobster_report/data/pizza_unit_tests.lobster
@@ -1,0 +1,159 @@
+{
+  "data": [
+    {
+      "tag": "python test_pizza_implementation.TestOvenController.test_start_cooking",
+      "location": {
+        "kind": "file",
+        "file": "test_pizza_implementation.py",
+        "line": 18,
+        "column": null
+      },
+      "name": "test_pizza_implementation.TestOvenController.test_start_cooking",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.oven_controller"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python test_pizza_implementation.TestOrderProcessor.test_create_customer",
+      "location": {
+        "kind": "file",
+        "file": "test_pizza_implementation.py",
+        "line": 33,
+        "column": null
+      },
+      "name": "test_pizza_implementation.TestOrderProcessor.test_create_customer",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.order_processor"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python test_pizza_implementation.TestOrderProcessor.test_place_order",
+      "location": {
+        "kind": "file",
+        "file": "test_pizza_implementation.py",
+        "line": 43,
+        "column": null
+      },
+      "name": "test_pizza_implementation.TestOrderProcessor.test_place_order",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.order_processor"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python test_pizza_implementation.TestPaymentGateway.test_process_payment_valid",
+      "location": {
+        "kind": "file",
+        "file": "test_pizza_implementation.py",
+        "line": 57,
+        "column": null
+      },
+      "name": "test_pizza_implementation.TestPaymentGateway.test_process_payment_valid",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.payment_gateway"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python test_pizza_implementation.TestInventoryManager.test_use_ingredients",
+      "location": {
+        "kind": "file",
+        "file": "test_pizza_implementation.py",
+        "line": 71,
+        "column": null
+      },
+      "name": "test_pizza_implementation.TestInventoryManager.test_use_ingredients",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.stock_manager"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python test_pizza_implementation.TestNotificationSystem.test_send_sms",
+      "location": {
+        "kind": "file",
+        "file": "test_pizza_implementation.py",
+        "line": 86,
+        "column": null
+      },
+      "name": "test_pizza_implementation.TestNotificationSystem.test_send_sms",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.notification_system"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python test_pizza_implementation.TestRecipeDatabase.test_get_recipe",
+      "location": {
+        "kind": "file",
+        "file": "test_pizza_implementation.py",
+        "line": 100,
+        "column": null
+      },
+      "name": "test_pizza_implementation.TestRecipeDatabase.test_get_recipe",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.recipe_database"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    },
+    {
+      "tag": "python test_pizza_implementation.TestRouteOptimizer.test_calculate_route_empty",
+      "location": {
+        "kind": "file",
+        "file": "test_pizza_implementation.py",
+        "line": 114,
+        "column": null
+      },
+      "name": "test_pizza_implementation.TestRouteOptimizer.test_calculate_route_empty",
+      "messages": [],
+      "just_up": [],
+      "just_down": [],
+      "just_global": [],
+      "refs": [
+        "req pizza_example.route_planner"
+      ],
+      "language": "Python",
+      "kind": "Method"
+    }
+  ],
+  "generator": "lobster_python",
+  "schema": "lobster-imp-trace",
+  "version": 3
+}
