+++ tests_system/lobster_report/lobster_report_system_test_case_base.py
@@ -1,6 +1,8 @@
 from pathlib import Path
-from .lobster_report_test_runner import LobsterReportTestRunner
-from ..system_test_case_base import SystemTestCaseBase
+from tests_system.lobster_report.lobster_report_test_runner import (
+    LobsterReportTestRunner,
+)
+from tests_system.system_test_case_base import SystemTestCaseBase
 
 
 class LobsterReportSystemTestCaseBase(SystemTestCaseBase):
