+++ tests_system/lobster_report/lobster_report_test_runner.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from pathlib import Path
 from typing import List, Optional
-from ..test_runner import TestRunner
+from tests_system.test_runner import TestRunner
 
 
 @dataclass
